changes to script

+++ cnp/data.py
@@ -166,9 +166,9 @@ class GPGenerator(DataGenerator):
         return np.squeeze(self.gp(x).sample())
 
     def log_like(self, x_context, y_context, x_target, y_target):
-        to_numpy = lambda x: x.squeeze().cpu().numpy()
-        post = self.gp.measure | (self.gp(to_numpy(x_context).astype(np.float64)), to_numpy(y_context).astype(np.float64))
-        return post(self.gp(to_numpy(x_target).astype(np.float64))).logpdf(to_numpy(y_target).astype(np.float64))
+        to_numpy = lambda x: x.squeeze().cpu().numpy().astype(np.float64)
+        post = self.gp.measure | (self.gp(to_numpy(x_context)), to_numpy(y_context))
+        return post(self.gp(to_numpy(x_target))).logpdf(to_numpy(y_target))
 
 
 class SawtoothGenerator(DataGenerator):
